requirements.txt
langchain-openai
numpy 
pandas
python-dotenv
langchain_community
langchain_core
langchain_groq
streamlit
streamlit_transformers_torch